pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pkg-auth"
version = "1.0.0"
description = "Clean-architecture auth core for multiple Python frameworks"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
    { name = "Fritill", email = "info@fritill.ae" }
]

# Runtime dependencies
dependencies = [
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "mypy>=1.11.0,<2.0.0",
    "types-requests>=2.32.0.0,<3.0.0.0",
    "types-PyJWT>=1.7.0,<2.0.0",
    "testcontainers[postgres,redis]>=4.0",
]

# ACL storage backends (v1.0)
acl-sqlalchemy = [
    "sqlalchemy>=2.0,<3.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
]
acl-django = [
    "django>=4.2",
    "psycopg[binary]>=3.1",
]

# Cache backends (v1.0)
cache-redis = [
    "redis>=5.0",
]

# Framework integrations
fastapi = [
    "fastapi>=0.115",
    "starlette>=0.37",
]
django = [
    "django>=4.2",
]
strawberry = [
    "strawberry-graphql>=0.255",
]

# Convenience meta-extra
all = [
    "fastapi>=0.115",
    "starlette>=0.37",
    "django>=4.2",
    "psycopg[binary]>=3.1",
    "strawberry-graphql>=0.255",
    "sqlalchemy>=2.0,<3.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "redis>=5.0",
]


[project.urls]
Homepage = "https://github.com/itqadem-apps/pkg_auth"

[project.scripts]
keycloak-init-client = "pkg_auth.admin.cli:main"

[tool.setuptools.packages.find]
where = ["src"]


[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pkg_auth.authentication.*"
strict = true
disallow_any_generics = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true

[[tool.mypy.overrides]]
module = "pkg_auth.authorization.*"
strict = true
disallow_any_generics = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true

[[tool.mypy.overrides]]
module = "pkg_auth.integrations.*"
strict = true
# Slightly relaxed: framework decorators (Django/Strawberry) sometimes need Any
disallow_any_expr = false
warn_return_any = false


[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-m 'not integration'"
markers = [
    "integration: tests requiring real Postgres or Redis (testcontainers)",
]
